Update Get-MTU.ps1
Applied fix to surprise bug where if MaxMTU was set below Interface capability and -Set was on, it would continue past the value due to internal fragmentation which resulted in the DF bit being applied to fragmented packets.

Changed "While" statement to reflect conditions the script is looking for.  Shortened script ~20 lines.
Added $LastLatency variable to reduce timeout of failed verification/retry.