Update Get-MTU.ps1
Added -Retry switch for inconsistent connections.
Reduced default MaxMTU to 1544 because anything higher can be defined and is also irregular on for regular users.
Noticed that if used as a start-up script, the "temporary fix" would always check for a higher value.  Not an intended feature... but not complaining.

Got the script to consistently measure under 15 ms, vice 1.5 seconds with original script that used ping.exe!  .NET is WAY quicker than Ping!